pyproject.toml
[tool.black]
line-length = 88
safe = true
skip-string-normalization = false
target_version = ['py38']

[tool.poetry]
name = "timeuntil"
version = "0.1.0"
description = ""
authors = ["Davide Sarra <davide.sarra@gmail.com>"]

[tool.poetry.dependencies]
python = "~=3.8"
pendulum = "^2.0.5"
Flask-Bootstrap = "^3.3.7"
Flask-WTF = "^0.14.3"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
